mana HUD overlay (text only) + add serverbound packet for spell switching

+++ src/main/java/com/teamsimplyrs/prismaarcanum/api/spell/spells/ignis/FireballSpell.java
@@ -23,7 +23,7 @@ public class FireballSpell extends AbstractSpell {
     public static final School school = School.Annihilation;
 
     public static final int tier = 1;
-    public static final float basicManaCost = 10f;
+    public static final int basicManaCost = 10;
     public static final float basicCooldown = 1f;
 
     public static final boolean hasEvolution = true;
